// forntEndApp/src/app/Models/Department.ts
export class Department {
    depId: number = 0;
    entId: number = 0;
    depCreatedBy: string = '';
    depCreatedDate: any;
    depModifiedBy: string = '';
    depModifiedDate: any;
    depStatus: boolean = true;
    depDescription: string = '';
    depName: string = '';
    depPhone: string = '';

}