// forntEndApp/src/app/Models/Employee.ts
export class EmployeeClass {
    empId: number = 0;
    empCreatedBy: string = '';
    empCreatedDate: any;
    empModifiedBy: string = '';
    empModifiedDate: any;
    empStatus: boolean = true;
    empAge: number = 0;
    empEmail: string = '';
    empName: string = '';
    empPosition: string = '';
    empSurname: string = '';

}